Move resource serializer to Query Builder
